export chat feedback in component + ui fixes
